Refactor OptionsUI into modules
